# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Mansis.Pos.Api | v1
  version: 1.0.0
servers:
  - url: http://localhost:5088/
paths:
...
  /api/v1/stock/movements:
    get:
      tags:
        - Stock
      parameters:
        - name: companyId
          in: query
          schema:
            type: string
            format: uuid
        - name: storeId
          in: query
          schema:
            type: string
            format: uuid
        - name: productId
          in: query
          schema:
            type: string
            format: uuid
        - name: movementType
          in: query
          schema:
            $ref: '#/components/schemas/StoreProductMovementType'
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: integer
            format: int32
            default: 1
        - name: pageSize
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type: integer
            format: int32
            default: 50
        - name: sort
          in: query
          schema:
            type: string
        - name: filter
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PagedResultOfStockMovementDto'
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResultOfStockMovementDto'
            text/json:
              schema:
                $ref: '#/components/schemas/PagedResultOfStockMovementDto'
  /api/v1/stock/stock-in:
    post:
      tags:
        - Stock
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockAdjustmentRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/StockAdjustmentRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/StockAdjustmentRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/StockMovementDto'
            application/json:
              schema:
                $ref: '#/components/schemas/StockMovementDto'
            text/json:
              schema:
                $ref: '#/components/schemas/StockMovementDto'
  /api/v1/stock/stock-out:
    post:
      tags:
        - Stock
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockAdjustmentRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/StockAdjustmentRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/StockAdjustmentRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/StockMovementDto'
            application/json:
              schema:
                $ref: '#/components/schemas/StockMovementDto'
            text/json:
              schema:
                $ref: '#/components/schemas/StockMovementDto'
  /api/v1/stock/destroy:
    post:
      tags:
        - Stock
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DestroyStockRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/DestroyStockRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/DestroyStockRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/StockMovementDto'
            application/json:
              schema:
                $ref: '#/components/schemas/StockMovementDto'
            text/json:
              schema:
                $ref: '#/components/schemas/StockMovementDto'
  /api/v1/stock/count:
    post:
      tags:
        - Stock
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockCountRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/StockCountRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/StockCountRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/StockMovementDto'
            application/json:
              schema:
                $ref: '#/components/schemas/StockMovementDto'
            text/json:
              schema:
                $ref: '#/components/schemas/StockMovementDto'
  /api/v1/stock/transfers:
    post:
      tags:
        - Stock
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransferRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateTransferRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateTransferRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/StoreProductTransferDto'
            application/json:
              schema:
                $ref: '#/components/schemas/StoreProductTransferDto'
            text/json:
              schema:
                $ref: '#/components/schemas/StoreProductTransferDto'
  /api/v1/stock/transfers/{id}/confirm:
    post:
      tags:
        - Stock
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmTransferRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/ConfirmTransferRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ConfirmTransferRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/StoreProductTransferDto'
            application/json:
              schema:
                $ref: '#/components/schemas/StoreProductTransferDto'
            text/json:
              schema:
                $ref: '#/components/schemas/StoreProductTransferDto'
  /api/v1/stock/transfers/{id}/receive:
    post:
      tags:
        - Stock
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiveTransferRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/ReceiveTransferRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ReceiveTransferRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/StoreProductTransferDto'
            application/json:
              schema:
                $ref: '#/components/schemas/StoreProductTransferDto'
            text/json:
              schema:
                $ref: '#/components/schemas/StoreProductTransferDto'
  /api/v1/stock/transfers/{id}/cancel:
    post:
      tags:
        - Stock
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelTransferRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CancelTransferRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CancelTransferRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/StoreProductTransferDto'
            application/json:
              schema:
                $ref: '#/components/schemas/StoreProductTransferDto'
            text/json:
              schema:
                $ref: '#/components/schemas/StoreProductTransferDto'
components:
  schemas:
...
    CancelTransferRequest:
      required:
        - companyId
        - userId
        - reason
      type: object
      properties:
        companyId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        reason:
          type: string
...
    ConfirmTransferRequest:
      required:
        - companyId
        - userId
      type: object
      properties:
        companyId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
...
    CreateTransferRequest:
      required:
        - companyId
        - userId
        - sourceStoreId
        - targetStoreId
        - lines
      type: object
      properties:
        companyId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        sourceStoreId:
          type: string
          format: uuid
        targetStoreId:
          type: string
          format: uuid
        lines:
          type: array
          items:
            $ref: '#/components/schemas/TransferLineWriteDto'
...
    DestroyStockRequest:
      required:
        - companyId
        - userId
        - storeId
        - productId
        - quantity
        - reason
      type: object
      properties:
        companyId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        storeId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        quantity:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: integer
          format: int32
        reason:
          type: string
...
    LedgerDirection:
      type: integer
    LedgerEntryState:
      type: integer
...
    PagedResultOfStockMovementDto:
      required:
        - items
        - page
        - pageSize
        - totalCount
        - totalPages
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StockMovementDto'
        page:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: integer
          format: int32
        pageSize:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: integer
          format: int32
        totalCount:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: integer
          format: int32
        totalPages:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: integer
          format: int32
...
    ProductTransferState:
      type: integer
    ProductUnitType:
      type: integer
...
    ReceiveTransferLineDto:
      required:
        - productId
        - receivedQuantity
      type: object
      properties:
        productId:
          type: string
          format: uuid
        receivedQuantity:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: integer
          format: int32
    ReceiveTransferRequest:
      required:
        - companyId
        - userId
        - lines
      type: object
      properties:
        companyId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        lines:
          type: array
          items:
            $ref: '#/components/schemas/ReceiveTransferLineDto'
...
    StockAdjustmentRequest:
      required:
        - companyId
        - userId
        - storeId
        - productId
        - quantity
        - description
      type: object
      properties:
        companyId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        storeId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        quantity:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: integer
          format: int32
        description:
          type: string
          nullable: true
    StockCountRequest:
      required:
        - companyId
        - userId
        - storeId
        - productId
        - countedQuantity
        - description
      type: object
      properties:
        companyId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        storeId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        countedQuantity:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: integer
          format: int32
        description:
          type: string
          nullable: true
    StockMovementDto:
      required:
        - id
        - companyId
        - storeId
        - productId
        - operationId
        - movementType
        - direction
        - quantity
        - state
        - reversalOfId
        - description
        - occurredAt
        - currentQuantity
      type: object
      properties:
        id:
          type: string
          format: uuid
        companyId:
          type: string
          format: uuid
        storeId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        operationId:
          type: string
          format: uuid
          nullable: true
        movementType:
          $ref: '#/components/schemas/StoreProductMovementType'
        direction:
          $ref: '#/components/schemas/LedgerDirection'
        quantity:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: integer
          format: int32
        state:
          $ref: '#/components/schemas/LedgerEntryState'
        reversalOfId:
          type: string
          format: uuid
          nullable: true
        description:
          type: string
          nullable: true
        occurredAt:
          type: string
          format: date-time
        currentQuantity:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: integer
          format: int32
...
    StoreProductMovementType:
      type: integer
    StoreProductTransferDto:
      required:
        - id
        - companyId
        - sourceStoreId
        - targetStoreId
        - requestedById
        - requestedTime
        - transferState
        - confirmedById
        - confirmedTime
        - receivedById
        - receivedTime
        - cancelledById
        - cancelledTime
        - cancelReason
        - transferDone
        - lines
      type: object
      properties:
        id:
          type: string
          format: uuid
        companyId:
          type: string
          format: uuid
        sourceStoreId:
          type: string
          format: uuid
        targetStoreId:
          type: string
          format: uuid
        requestedById:
          type: string
          format: uuid
        requestedTime:
          type: string
          format: date-time
        transferState:
          $ref: '#/components/schemas/ProductTransferState'
        confirmedById:
          type: string
          format: uuid
          nullable: true
        confirmedTime:
          type: string
          format: date-time
          nullable: true
        receivedById:
          type: string
          format: uuid
          nullable: true
        receivedTime:
          type: string
          format: date-time
          nullable: true
        cancelledById:
          type: string
          format: uuid
          nullable: true
        cancelledTime:
          type: string
          format: date-time
          nullable: true
        cancelReason:
          type: string
          nullable: true
        transferDone:
          type: boolean
          nullable: true
        lines:
          type: array
          items:
            $ref: '#/components/schemas/StoreProductTransferLineDto'
    StoreProductTransferLineDto:
      required:
        - id
        - productId
        - quantity
        - receivedQuantity
        - unit
        - unitPrice
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        quantity:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: integer
          format: int32
        receivedQuantity:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: integer
          format: int32
          nullable: true
        unit:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/ProductUnitType'
        unitPrice:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type: number
          format: double
          nullable: true
...
    TransferLineWriteDto:
      required:
        - productId
        - quantity
        - unit
        - unitPrice
      type: object
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          pattern: ^-?(?:0|[1-9]\d*)$
          type: integer
          format: int32
        unit:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/ProductUnitType'
        unitPrice:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type: number
          format: double
          nullable: true
...